python_api/requirements.txt
cassandra-driver>=3.1
tornado>=4.5.1
